chore: memo 폴더 생성 후 컴포넌트 경로 수정
